python/practise/Python前后端分离开发Vue+DRF实战/MxShop/requirements.txt
django==2.0
pillow
mysqlclient
djangorestframework-jwt
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
django-import-export>=0.5.1
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven